Split the gum renders into Pre and Post-based events.
Health now renders in the Pre event (as does Mantle), allowing us to
catch cancelled events; a Post event is never fired for Health if the
Pre event was cancelled, resulting in the code never being called with
Mantle installed.

The event is kept cancelled to prevent Minecraft from performing
additional shenanigans.

Hunger is left on its own in the Post event.

Both are set up to receive canceled events in the event that another
event handler cancels the event.